chore: rename yarn-or-npm -> package-manager
