+++ packages/utils/core-utils/src/index.ts
@@ -1,3 +1,3 @@
 export * from './rebuild';
 export * from './electron-version';
-export * from './yarn-or-npm';
+export * from './package-manager';
